+++ menu-module/build.gradle
@@ -1,19 +1,8 @@
-plugins {
-    id 'java'
-}
-
-group 'camp.nextstep.edu'
-version '0.0.1-SNAPSHOT'
-
-repositories {
-    mavenCentral()
-}
-
 dependencies {
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.0'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.7.0'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    testImplementation 'com.h2database:h2'
 }
 
-test {
-    useJUnitPlatform()
+jar {
+    enabled = true
 }
