+++ menu-module/src/main/java/kitchenpos/product/domain/Product.java
@@ -2,6 +2,7 @@ package kitchenpos.product.domain;
 
 import kitchenpos.global.BaseTimeEntity;
 
+import javax.persistence.*;
 import java.math.BigDecimal;
 
 @Entity
