--- menu-module/src/main/resources/db.migration/V1__Initialize_project_tables.sql
@@ -1,40 +1,0 @@
-CREATE TABLE menu (
-    id BIGINT(20) NOT NULL AUTO_INCREMENT,
-    name VARCHAR(255) NOT NULL,
-    price DECIMAL(19, 2) NOT NULL,
-    menu_group_id BIGINT(20) NOT NULL,
-    PRIMARY KEY (id)
-);
-
-CREATE TABLE menu_group (
-    id BIGINT(20) NOT NULL AUTO_INCREMENT,
-    name VARCHAR(255) NOT NULL,
-    PRIMARY KEY (id)
-);
-
-CREATE TABLE menu_product (
-    seq BIGINT(20) NOT NULL AUTO_INCREMENT,
-    menu_id BIGINT(20) NOT NULL,
-    product_id BIGINT(20) NOT NULL,
-    quantity BIGINT(20) NOT NULL,
-    PRIMARY KEY (seq)
-);
-
-CREATE TABLE product (
-    id BIGINT(20) NOT NULL AUTO_INCREMENT,
-    name VARCHAR(255) NOT NULL,
-    price DECIMAL(19, 2) NOT NULL,
-    PRIMARY KEY (id)
-);
-
-ALTER TABLE menu
-    ADD CONSTRAINT fk_menu_menu_group
-        FOREIGN KEY (menu_group_id) REFERENCES menu_group (id);
-
-ALTER TABLE menu_product
-    ADD CONSTRAINT fk_menu_product_menu
-        FOREIGN KEY (menu_id) REFERENCES menu (id);
-
-ALTER TABLE menu_product
-    ADD CONSTRAINT fk_menu_product_product
-        FOREIGN KEY (product_id) REFERENCES product (id);
