--- menu-module/src/main/resources/db.migration/V3__create_modify_date.sql
@@ -1,12 +1,0 @@
-ALTER TABLE menu ADD created_date DATETIME;
-ALTER TABLE menu_group ADD created_date DATETIME;
-ALTER TABLE menu_product ADD created_date DATETIME;
-ALTER TABLE product ADD created_date DATETIME;
-ALTER TABLE menu ADD modified_date DATETIME;
-ALTER TABLE menu_group ADD modified_date DATETIME;
-ALTER TABLE menu_product ADD modified_date DATETIME;
-ALTER TABLE product ADD modified_date DATETIME;
-
-
-
-
