+++ menu-module/src/test/java/kitchenpos/AcceptanceTest.java
@@ -1,0 +1,59 @@
+package kitchenpos;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.http.MediaType;
+
+import java.sql.SQLException;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class AcceptanceTest {
+
+    @Autowired
+    private DatabaseCleanup dataBaseCleanUp;
+
+    @LocalServerPort
+    int port;
+
+    @BeforeEach
+    public void setUp() {
+        RestAssured.port = port;
+    }
+
+    @AfterEach
+    public void after() throws SQLException {
+        dataBaseCleanUp.truncate();
+    }
+
+    protected ExtractableResponse<Response> post(final String url, final Object body) {
+        return RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(body)
+                .when().post(url)
+                .then().log().all()
+                .extract();
+    }
+
+    protected ExtractableResponse<Response> put(final String url, final Object body, final Object pathVariable) {
+        return RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(body)
+                .when().put(url, pathVariable)
+                .then().log().all()
+                .extract();
+    }
+
+    protected ExtractableResponse<Response> delete(final String url, final Object pathVariable) {
+        return RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when().delete(url, pathVariable)
+                .then().log().all()
+                .extract();
+    }
+}
