+++ menu-module/src/test/java/kitchenpos/MenuModuleApplicationTests.java
@@ -1,0 +1,7 @@
+package kitchenpos;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MenuModuleApplicationTests {
+}
