--- order-module/src/main/resources/application.yml
@@ -1,31 +1,0 @@
-spring:
-  profiles:
-    active: dev
----
-spring:
-  config:
-    activate:
-      on-profile: prod
----
-spring:
-  config:
-    activate:
-      on-profile: dev
-  h2:
-    console:
-      enabled: true
-  jpa:
-    hibernate:
-      dialect: org.hibernate.dialect.H2Dialect
-      ddl-auto: validate
-    show-sql: true
-    properties:
-      hibernate:
-        format_sql: true
-logging:
-  level:
-    org:
-      hibernate:
-        type:
-          descriptor:
-            sql: trace
