+++ order-module/src/test/java/kitchenpos/OrderModuleApplicationTests.java
@@ -1,0 +1,9 @@
+package kitchenpos;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@EnableJpaAuditing
+@SpringBootApplication
+public class OrderModuleApplicationTests {
+}
